# Conflicts:
#	src/main/java/za/ac/cput/controller/ContactController.java
#	src/main/java/za/ac/cput/domain/Contact.java
#	src/main/java/za/ac/cput/service/ContactService.java
#	src/main/java/za/ac/cput/service/IContactService.java

+++ src/main/java/za/ac/cput/service/ICartItemService.java
@@ -11,6 +11,6 @@ Date: 03 August 2025
 
 import java.util.List;
 
-public interface ICartItemService extends IService<CartItem,Integer>{
+ public interface ICartItemService extends IService<CartItem,Long>{
     List<CartItem> getAll();
 }
